refactor: differenciate between modes and add seeds
