+++ lib/commands/seed.js
@@ -6,7 +6,7 @@ var _assert = require('./helper/assert');
 
 function executeSeed (internals, config, callback) {
   var index = require('../../connect');
-  var Seeder = require('../seeder.js');
+  var Seeder = require('../walker.js');
 
   if (internals.argv._.length > 0) {
     internals.argv.destination = internals.argv._.shift().toString();
@@ -15,7 +15,8 @@ function executeSeed (internals, config, callback) {
   index.connect(
     {
       config: config.getCurrent().settings,
-      internals: internals
+      internals: internals,
+      prefix: 'seed'
     },
     Seeder,
     function (err, seeder) {
