+++ lib/commands/undo-seed.js
@@ -7,7 +7,7 @@ var _assert = require('./helper/assert');
 
 function executeUndoSeed (internals, config, callback) {
   var index = require('./connect');
-  var Seeder = require('./lib/seeder.js');
+  var Seeder = require('./lib/walker.js');
 
   if (!internals.argv.count) {
     log.info('Defaulting to running 1 down seed.');
@@ -21,7 +21,8 @@ function executeUndoSeed (internals, config, callback) {
   index.connect(
     {
       config: config.getCurrent().settings,
-      internals: internals
+      internals: internals,
+      prefix: 'seed'
     },
     Seeder,
     function (err, seeder) {
