+++ lib/walker.js
@@ -29,7 +29,11 @@ const Walker = function (driver, directory, mode, intern, prefix) {
   Promise.promisifyAll(this._driver);
   this.directory = directory;
   this.internals = intern;
-  this.prefix = prefix;
+  this.mode = mode;
+
+  if (!this.mode) this.prefix = `static-${prefix}`;
+  else this.prefix = prefix;
+
   this.title = `[${prefix}] `;
 
   // keep it until we decide how we do the cross linking
